shop_tg_app/requirements.txt
asgiref==3.8.1
Django==5.2.1
django-js-asset==3.1.2
django-mptt==0.17.0
dotenv==0.9.9
gunicorn==23.0.0
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
sqlparse==0.5.3
gevent>=1.4